release 3.2, Jan 4, 1980
SCCS-vsn: usr.bin/ex/ex.c 2.1
SCCS-vsn: usr.bin/ex/ex.h 2.1
SCCS-vsn: usr.bin/ex/ex_addr.c 2.1
SCCS-vsn: usr.bin/ex/ex_argv.h 2.1
SCCS-vsn: usr.bin/ex/ex_cmds.c 2.1
SCCS-vsn: usr.bin/ex/ex_cmds2.c 2.1
SCCS-vsn: usr.bin/ex/ex_cmdsub.c 2.1
SCCS-vsn: usr.bin/ex/ex_data.c 2.1
SCCS-vsn: usr.bin/ex/ex_get.c 2.1
SCCS-vsn: usr.bin/ex/ex_io.c 2.1
SCCS-vsn: usr.bin/ex/ex_put.c 2.1
SCCS-vsn: usr.bin/ex/ex_re.c 2.1
SCCS-vsn: usr.bin/ex/ex_re.h 2.1
SCCS-vsn: usr.bin/ex/ex_set.c 2.1
SCCS-vsn: usr.bin/ex/ex_subr.c 2.1
SCCS-vsn: usr.bin/ex/ex_temp.c 2.1
SCCS-vsn: usr.bin/ex/ex_temp.h 2.1
SCCS-vsn: usr.bin/ex/ex_tty.c 2.1
SCCS-vsn: usr.bin/ex/ex_tty.h 2.1
SCCS-vsn: usr.bin/ex/ex_tune.h 2.1
SCCS-vsn: usr.bin/ex/ex_v.c 2.1
SCCS-vsn: usr.bin/ex/ex_vadj.c 2.1
SCCS-vsn: usr.bin/ex/ex_vars.h 2.1
SCCS-vsn: usr.bin/ex/ex_vget.c 2.1
SCCS-vsn: usr.bin/ex/ex_vis.h 2.1
SCCS-vsn: usr.bin/ex/ex_vmain.c 2.1
SCCS-vsn: usr.bin/ex/ex_voper.c 2.1
SCCS-vsn: usr.bin/ex/ex_vops.c 2.1
SCCS-vsn: usr.bin/ex/ex_vops2.c 2.1
SCCS-vsn: usr.bin/ex/ex_vops3.c 2.1
SCCS-vsn: usr.bin/ex/ex_vput.c 2.1
SCCS-vsn: usr.bin/ex/ex_vwind.c 2.1
SCCS-vsn: usr.bin/ex/ex3.7preserve/ex3.7preserve.c 2.1
SCCS-vsn: usr.bin/ex/ex3.7recover/ex3.7recover.c 2.1
SCCS-vsn: usr.bin/ex/Makefile 2.1
SCCS-vsn: usr.bin/ex/OTHER/makefile.70 2.1
SCCS-vsn: usr.bin/ex/makeoptions 2.1
SCCS-vsn: usr.bin/ex/printf.c 2.1

+++ usr/src/usr.bin/ex/ex_argv.h
@@ -14,7 +14,7 @@ short	argc;
 short	argc0;
 short	morargc;		/* Used with "More files to edit..." */
 
-short	firstln;		/* From +lineno */
+int	firstln;		/* From +lineno */
 char	*firstpat;		/* From +/pat	*/
 
 /* Yech... */
